request #14844: JS error when going back to a campaign
1. Load a campaign,
2. Using the breadcrumb go back to list of campaigns,
3. Select the same campaign as before.
=> you should not have any errors in the console

This was due to the fact that the content of the campaign was already
loaded, and we tried to apply a filter on a non fully filled $scope.
There is no error on fist load because the campaign is not loaded and we
don't need to apply the filter yet.

Now the $scope is set before the load of the campaign.

Change-Id: I2f114ca03368982b06e7ded22ba1f75b3a93a492