request #14867: Introduce embed as a new plugin
Here comes a new plugin.

Install and activate the plugin "Embed".
Go to an artifact. Insert a public Figma url:
* in a text field
* in a string field
* in a follow-up comment

Submit the artifact. Go to the artifact.

Under the url a live preview of the service url is displayed.

This patch only deals with Figma, based on their recommendation[0].
Other services will be added in dedicated commits.

This takes another approach than previous implementation
that was suffering from various security issues[1].

[0] https://www.figma.com/developers/embed
[1] git #tuleap/stable/ef8ada3d69af2a23aa3f2484cd24c6be3bd7e1c7

Change-Id: I7c30d43e9b38b6d3b8629cf453893951388d5ef9