request #14867: Introduce embed-js as a new plugin
Here comes a new plugin.

Install and activate the plugin "Embed".
Go to an artifact. Insert a public codepen, soundcloud, giphy, or vimeo
url:
* in a text field
* in a follow-up comment

Submit the artifact. Go to the artifact.

Under the url a live preview of the service url is displayed (a player,
the gif image, …).

See README of the plugin for more details.

Change-Id: I5920972be348173dd5d81b6ffecb7aeaf5f429d9