Have a dedicated object to build string field changes
This is part of story #14766 Have an import from jira structure.

This is a refactoring, no functional changes expected here.
The aim of this refactoring is to centralize in a dedicated object the
building of the XML field change node regarding string changes,
for both Tuleap export and Jira import process.

Change-Id: I03a3558b9a9aea5572ac7e59c6910365c512a4a0