Have a dedicated object to build text field changes
This is part of story #14766 Have an import from jira structure.

This is a refactoring, no functional changes expected here.
The aim of this refactoring is to centralize in a dedicated object the
building of the XML field change node regarding text changes,
for both Tuleap export and Jira import process

Change-Id: I1200f007d5554012bc20314078b3a4791e4c51b5