Have a dedicated object to build float field changes
This is part of story #14766 Have an import from jira structure.

This is a refactoring, no functional changes expected here.
The aim of this refactoring is to centralize in a dedicated object the
building of the XML field change node regarding float changes,
for both Tuleap export and Jira import process

Change-Id: Ib13c8b8772989cc98dd0870f13f7c98bcc7b2be8