Remove last custom usage of logger in Core + Internal Plugins
Introduce usage of a trait for what traits are good for: copy/paste.

There are still a few usage of BackendLogger in External Plugins that
will go away soon.

Be careful, in order to test you need to re-generate autoload because
src/common/log is not PSR4.

Part of request #14864 Centralize logs in syslog & log forwarding

Change-Id: I4f4f5914319197bff17825c3c6171f22003201f1