Inject dl limits in Vue app
As site administrator go to admin » Document » File download limit and
change settings.
Open docman with the new UI. Nothing visible change, but if you inspect
the Vue's store (with Vue.js Devtools for example) then you will see that
the previous settings are in the state.

Part of story #72289: download folder as zip

Change-Id: I75c32a68f727c1ecad168f4b2bd7a88d69bf3180