HTML fields should be imported as HTML format
In Jira the raw value of HTML fields is wiki style format. Since Tuleap
is not able to render this format, we should store the rendered value
instead.

This is done by using Expansion in Jira REST API[0].

To test enter some HTML in a text field, and some wiki markup in a html
field like "Description".
Import the issue in Tuleap.
=> the text field is stored as text format, the HTML is not interpreted.
=> the Description is stored as HTML format.

[0] https://docs.atlassian.com/software/jira/docs/api/REST/5.2.11/#expansion

Part of story #14766: Have an import from jira structure

Change-Id: I2697630f8d72d5678a0982276d97f099d60559bd