Forward application logs to a Graylog server
Leveraging monolog adapters it's now easy to forward logs to a
graylog server.

There are new variables to configure (config-set):
- graylog2_server
- graylog2_port
- graylog2_ssl

And finally:
tuleap config-set sys_logger graylog2

Note: by default it will log using TCP transport, that might not
be scalable enough on high traffic sites but this is mandatory for
out of the box usage of TLS.
In case of overloading, we should log in UDP and have something
in between to make the proxy.

In order to test, if you know admins that have a working graylog2
server, request them to create a new input for you ;)

Part of request #14878 Forward logs to Graylog server

Change-Id: I726fbe093e415c72da31027d3ccea32d193d1a93