Clean obsolete Jira import
We should not keep forever the pending jira import. One day to process
the import is considered enough. The user who initiated the import is
notified about the cancellation.

Start a Jira import. Once created there is a feedback info telling us
that an import is pending. Update in database the created_on to simulate
an import initiated 1 or 2 days ago. Then run `tuleap daily-job`.

* obsolete Jira imports are removed from database
* people are notified

This is hidden behind a feature flag. Please issue the following command
in order to use it:

  tuleap config-set asynchronous_jira_creation 1

Reminder: with this setting, the tracker **will not be created**, it is
here only for testing purpose.

Part of story #14766: Have an import from jira structure

Change-Id: I08d72ec863865ee6e9c8fa029b4c9cb1acc4fda0