Jira is imported asynchronously
This is hidden behind a feature flag. Please issue the following command
in order to use it:

  tuleap config-set asynchronous_jira_creation 1

Once created in the interface, an event will be processed and the Jira
issues will be imported.
* In case of error (ex: wrong credentials, xml parse, …) the user will
receive an email about this error.
* In case of success, the user will receive an email with the link to
the new tracker.

In any case, the pending import is removed, forcing the user to start
over in case of error. If this design choice appears to be too
cumbersome, then we could adjust this later.

You must have a working Redis in your stack[0]

Part of story #14766: Have an import from jira structure

[0] https://docs.tuleap.org/developer-guide/back-end/redis.html

Change-Id: I6b8bf8fd5e11ffd24c6fc86036952c9e39de32e8