Corrupted file break the process
part of: story #12404 download folder as zip

How to test:
- Create a file (File or Embedded_File)
- Delete its current_version in the plugin_docman_version table
- In the UI, your file has a [CORRUPTED] badge
- Download the archive

--> No error
--> The corrupted file is logged in codendi_syslog
--> The archhive is well generated, does not contain the corrupted file

Although this file does not appear in the archive, we should probably
create a file in the archive with the same name and the following text: "this file is
corrupted, it couldn't be append to the archive since it has no
current version". May be done in another commit.

Change-Id: I16ced99d3526d62b2dbacec5c0c015726a7386b8