Add a text file that lists problematic files in the archive
Part of: story #12404 download folder as zip

How to test:
- Edit ZipStreamFolderFilesVisitor::addFileToArchive and append any
string to the $fs_path (in order to trigger file not found errors)
- Download a folder as a zip archive
- The zip archive should contain a file named 'TULEAP_ERRORS.txt'
containing a translated text and the paths of files that could not be
downloaded.

When there is no error, the TULEAP_ERRORS.txt file is not created.

Change-Id: I155e5477ec018b30253bc5fafcb76e5df2380fcc