Show confirm modal
part of: story #12404 download folder as zip

A confirmation modal is needed to tell the user what will be contained
in the archive.

It also warns about the existence of TULEAP_ERRORS.txt when a file could
not have been added to the archive. This warning has also been added to
the warning modal since the download is not prevented.

I took the opportunity to do a micro-refactoring of the way we trigger
the download. Now, we don't set window.location with the href of the
archive anymore, we define confirm buttons as links whose hrefs point to
it.

How to test:
- Pick up a light folder, click on [Download as a zip]
--> The confirm modal opens

- Click cancel
--> It closes the modal

- Click Download
--> The download starts

- Same behavior with the warning modal

Change-Id: I092d19ec732021b509cba6c4fac23b884e4a87e9