Error during the retrieval of the folder's size shows a modal
Part of story #12404 download folder as zip

Instead of hiding the entire hierarchy and leaving people unable to
continue their work (previous error handling), this contribution adds a
"global" error modal. It is "global" because there is already an "error
modal" state. It opens a modal that shows a generic error message and
has an anchor to show the error's details to a user. In our case, it
will be the http error code and message. We use the exact same pattern
in Taskboard.

How to test:
- after loading the document tree, simulate an error in
DocmanItemsResource.php (throw a RestException)
- alternatively, simulate being offline in your browser dev tools.
- When you click on "Download as a zip", you should see the error modal
pop up.

Change-Id: I8eb0115d3c2d8f1e50704df0e52b821d37af7359