request #14899 gl-membership script must be working for gitolite system user
The current call to gl-membership script by the gitolite user throws a fatal error.

A new sudo rule as been added to be sure that the gitolite user is able
to receive the good result for the call to gl-membership script.

To test it, you have to manually deploy the new sudoers and the
gl-membership.pl files and try to do git operations.
They must be allowed or denied as expected.

Change-Id: I9455b37d114caa4292a7d77639932443be9dff37