Fix gl-membership.pl on centos6
This fix a regression introduced by cfaf02210d.

In this commit we refactored the way database.inc was loaded and guard was
added to ensure file was reachable to produce a nice error message instead
of a weird stack trace.

However, the check done with file_exists doesnt work on centos6 / perl setuid
because there bugs in file_exists implementation in this context.
The workaround is to use is_file that does the same job but without the bug.

The problem doesn't arise on centos7 because it uses sudo which is a better
way to solve the problem. I choose not to go the sudo way in order to stay
as close as possible of the original code and avoid painful upgrades.

Change-Id: I1acfc7989816d79d8a94cff98b4d757c0fdbfba8