Remove download confirmation modal
Part of story #12404 download folder as zip

Since the modal is always opened, it is likely that people will click it
without reading when a problem does arise and there is actually a
TULEAP_ERRORS.txt file. We will instead send an email (in another
commit).

How to test:
- make tests_cypress should pass
- when you download a zip folder with size inferior to the "warning
limit", there won't be any modal.
- Modals are still shown for "warning limit", "OSX warning" and "Max
size limit" cases.

Note:
This solution brings back the Chromium warning:

Resource interpreted as Document but transferred with MIME type
application/x-zip

But it can't be worked around. If we put the download URI as href in the
link, people will be able to bypass the "size exceeded" modal with
ctrl+click.

Change-Id: I04758efcc8a6a218a8435234112ef18b10791b96