Allow download of big folders to finish its execution
Part of story #12404 download folder as zip

This removes the PHP max execution time limit on the "Download as zip"
controller. Keeping it will abruptly stop the download while it is not
finished, after 30 seconds of execution.
As site administrator MUST choose a max file size limit and it is set by
default to 2GB and it is checked and enforced in the controller, the
file size becomes the new limiting factor that will prevent the
controller from running indefinitely. We can therefore remove the max
execution limit.

How to test:
- Either change your network settings to have very low download speed
and/or try to download a very big archive (on my setup it was above
4GB). The download should not stop before it is complete.

Change-Id: Ie13a8ee1753d4b3fc385284e1d1b205b983ccdce