request #14920: Remove create from XML from legacy creation page
Creation of trackers from XML files is available via the new creation
interface. There is no need to keep it in the legacy one.

Go trackers » create new tracker » switch to legacy interface.

There should be only the following options:

* From template
* From TV3 (only in projects with TV3)

Note: if projects does not have TV3, then there is an orphan radio
button. It is not planned to fix this since the interface will be
removed soon.

Change-Id: If751123815bc571cb7f90fd2e02c2c20e13e1fc4