Display user info as tooltip on @mentions
If you @mention someone in a wiki page or in an artifact field, then
when you hover the link, there should be a tooltip displaying basic
information about the user.

This works both on FlamingParrot and BurningParrot pages.
For BurningParrot, you can add a renderer as a widget on project
dashboard, with the artifact fields as a column.

Part of request #14905: Have sexier cross reference tooltip

Change-Id: Ic18ce86e81ba5ad6c35f7191bc8c6a9b8d7053f5