Fix display of mini widgets in Project Milestones
This is part of request #14877 Fix project milestones widget UI

You have access to a figma in the request to see what expected in this patch.

The mini widgets are: description, tracker badges, charts (burnup, burndown, ttm)

If your screen size is below 1800px, the display change because there is not enough space to display as in the figma.

How to test:
	- Have some releases
	- Have some mini widgets
		=> The display is like in the figma

	- Have a long tracker's label
                => There is an ellipsis with a tooltip

Change-Id: I1b634869eb3e906fcb3428676d15f5fc0de55023