PHP unit tests can be executed without failures when /etc/pki/ca-trust/extracted/pem/tls-ca-bundle.pem does not exist
Unit tests introduced in 05664bad81d035494237d31c8f1343131e20d2dc have
considered the tests were always executed on a CentOS7+/Fedora based
systems. Other systems have different paths for the CA bundle.

The situation is annoying in a dev environnement that is not Fedora
based as you get failures that are "normal". Also, it should be expected
to execute the test suite with success in environnements similar to production
and (unfortunately) there are still Tuleap instances running on CentOS 6.

In any cases, a setting was made to find the path of this file.
Hardcoding it in the test cases should not be required.

Part of story #14768 have a tuleap enterprise edition docker image production ready

Change-Id: Ie0478d3b26e69345ab136dc3403eede98192bb41