No legacy tracker creation page for projects without tv3
If the project is using tv3, then there is a link to access the legacy
UI.

If the project is not using tv3, then the link is not present.

Note: if the user has bookmarked the link, then it is still accessible,
even if there isn't any tv3 in the project.

A dedicated changeset will remove the create from template section in
this legacy interface.

request #14932: Legacy tracker creation page should only be there for TV3 migration

Change-Id: I1b4ffa65b7cc2756e0c59e3365c7d828ceafb413