Legacy tracker creation page deals only with tv3
It is not anymore possible to duplicate a tracker from another one with
the legacy interface. Users must use the new one instead.

If users bookmarked the page and there isn't any tv3 in the project,
then they are automatically redirected to the new interface.

request #14932: Legacy tracker creation page should only be there for TV3 migration

Change-Id: I8c4b551902917db35243e1adf5608f5b817bffec