Introduce new TTM PaneInfo
Go to AgileDashboard » pv2
=> on a milestone there is a new link "✓ Tests"

Click on the overview or taskboard link of this milestone
=> there is a new tab "Tests" next to the legacy "Test campaigns 🡕"

"✓ Tests" and "Tests" links to the legacy page (angular view in ttm).
A dedicated page will be added in a dedicated commit.

Note: the order of links in pv2 does not match the one in milestone
tabs. Furthermore the tabs to display is hard coded in pv2 and
references other plugins which is not necessarily a good thing. This
will be handled in a dedicated commit.

Part of story #14881: Transform “Test Management” pane in “Integrated pane” Test

Change-Id: I55ffd0fa14e5c9bc663e73defb3cb9c7944bc3dd