Refactoring: introduce a dedicated object to deal with changeset with data
This is part of story #14925 Import Jira issue comments

No functional changes expected here.
The same kind of object will be created to deal with comments.
(comments and issue data are 2 different things in Jira that will be
converted into changesets)

ArtiactXMLExporter is quite big. Moving this responsibility into a
dedicated object will ease the readability and the maintainability.

Change-Id: Ibb7c9a0cfef6f9da1722e72cd77fa5b2438411d3