Fix order of milestone links in pv2
Until now there was two different events to collect the list of pane for
a milestone. One for the REST route (to expose "additional_panes") and
another one for tabs + breadcrumb dropdown.

A precedent work had been done to ensure a suitable taskboard tab
position[0] for one of the event. However the other event didn't need to
be updated. Since the introduction of "Tests" icon in pv2, it is not
anymore the case and we need to have the same order everywhere.

The event for the REST route is therefore terminated so that we have the
order logic in one place only. In order to save resources and
rainforest, the building of active pane (that is irrelevant in REST
context) is deferred through a closure.

[0] git #tuleap/stable/ef8e58bb223378df760b72800d1b1a4f90c991fc

Part of story #14881: Transform “Test Management” pane in “Integrated pane” Test

Change-Id: I8066687ab06eba365a36ccb737ab1f73d84666d6