request #14936: The zip archive of a document folder should not be indicated as public information
As the zip archive of a folder takes into account the permissions it
cannot be put in a public cache without potentially leaking restricted
documents.

In order to fix the issue and to have a consistent behavior when
downloading large files from a Tuleap instance (without duplication
code), the controller has been transformed into something that can
support PSR-15 middleware and PSR-7 interfaces.

Change-Id: I6dd607c9624763a4c7e0276a1e14a87e6359b29b