Create inital changeset
This is part of story #14927 Import issue history

The ail of this patch is to recreate the inital state of the issue (the
first changeset). To do this, we retrieved the last state of the issue,
and then we do a diff with all the changelog to recreate at best this
initial state.

For now, 2 changesets are created by imported artifacts :

* the initial changeset with creation date
* the last changeset (copy of the current state) with last update date

The full changelog will be imported in next commits.

Change-Id: I4b5a715ad1319bb8cfcf8e40c87aacd4fe5fc48b