request #14954 Extract common configuration to transpile node_modules
How to test:
- Rebuild AgileDashboard, Taskboard and TestManagement
- Kanban and TestManagement should run on IE11 without crashing.
- Taskboard should run on "Chromium Edge" without crashing. Edge based
on EdgeHTML is no longer supported.

Per our babel configuration, we only support the last 2 versions of
Edge. Our documentation is even stricter, stating that we support the
latest version [0].

Edge (based on EdgeHTML) crashed on an Object Spread in "p-retry", one
of our dependencies. Edge (based on Chromium) supports it since version
79 (see [1]), and version 83 has just been released in May 2020. Even if
we kept this transpilation, the rest of the bundle contains Object
Spreads (from our own source code). We can drop it.

[0] https://docs.tuleap.org/user-guide/misc.html#user-supported-browsers
[1]
https://caniuse.com/#feat=mdn-javascript_operators_spread_spread_in_object_literals

Change-Id: Ie277019b10fb5b23c734570562044b98a371483f