Display list of campaigns
Go to the test plan of a milestone.

If the milestone has campaigns (open or closed), they are displayed.
Else the list is empty.

While the list is loading, skeletons are displayed.

This commit is intended to introduce vuex store and basic retrieval of
campaigns. As such, and to not clutter much the review, no polish has
been applied:
- "Test campaigns" title should be black
- Empty state should be displayed if the list is empty
- Error state should be displayed in case of error while loading the
list.
- The card for each campaign should be styled.
- The list is not ordered.

All this will be done later, in a dedicated changeset.

Part of story #14881: Transform “Test Management” pane in “Integrated pane” Test

Change-Id: Iccaad14382d9043e4cf3234bba1b2efcefab6ca1