Fix TTM chart label
This is part of request #14877 Fix project milestones widget UI

To make the widget more compact, now the TTM chart is moved higher.
Before this patch, the label was too far from the chart.

Now, if you hover a slice of the chart and the number goes above the label,
there is a white shadow all around the number to better see it.

No expected changes.

Change-Id: Ib6eedd8d9d804201f82dc4a24b9a508296be70fe