Refactoring: Introducing a collection of Snapshot Objects
This is part of story #14927 Import issue history

The aim of these refactorings is to ease the readability
and the maintainability of the whole feature.

Currently, we only have two states (initial and current). In next
commits, there will be multiple states based on history. Having this
collection will ease the manipulation of states and their XML export.

In addition, some classes have been renamed or splitted to have
more epxlicit and meaningful names and responsibilities.

Change-Id: Ic0f6f0f9651b01a6496cd9ad49f416688751548f