Add a helper to run unit tests with PHP 8.0
An new target is added in the Makefile to let developpers run the unit
tests with PHP 8.0. This target is formatted like the existing
`tests-rest` or `tests-soap` targets.

To play the unit tests with PHP 8.0: `make tests-unit-php PHP_VERSION=80`.
No fix has been done to the code yet so some errors are expected. Also
note that Mockery currently raises deprecation notices wich does not
play well with the default of the test suite error_reporting=-1.

Closes request #15016: Allow developers to run unit tests with PHP 8.0

Change-Id: I30a53c42e688271cf34ee345dd6b6bd7d0f91c80