Shorten the TTM breadcrumb for a milestone
Go to test management service
=> Breadcrumb should be [ Campaigns >

Go to a campaign
=> Breadcrumb should be [ Campaigns > Selected campaign >

Go to the test plan of a milestone, and select a campaign.
=> Breadcrumb should be [ Milestone title > Selected campaign >

Note: I'm not really proud of the current proposal which is not really
angularjs-ish. The whole routing (or the whole app) should be burned in
hell and rewritten anyway, therefore I don't think we should spend much
time on it right now.

Part of story #14882: Create new campaign in new “Test” screen

Change-Id: I16e0707d373ff78437364348868b4744ca1e71d2