Move UGroupLiteralizer into the Tuleap\Project namespace
This contribution deals with the move of the UGroupLiteralizer to make
it PSR-4 since it has been a topic of discussions in the reviews of
01cb08c83ee722adafc98ab369324ac7dcee9af7.

As the file has been moved, it does not match the existing issues
allowed in the Psalm's baseline. Minor fixes have been done to make
Psalm happy.

This change is part of story #14019: automate jenkins job creation & run

Change-Id: I63cb6bf5522f5447eb8a027308607d405fbe7130