Mark some .tab entries as overridable
Some existing users have customized some .tab entries and we should not
loose those customization during the tab2gettext automatisation.

.tab entries that are removed are not anymore used.

There is also a refactoring work on the global search, that used in the
past a selectbox to let the user choose in which part of the application
the search should be performed. Since this selectbox is not anymore
used, we can remove the corresponding labels.

Part of request #14979: Identifiy overridable strings with getOverridableText()

Change-Id: I5c5ea0953c52d1c74ab87e2a685522f6ca16cee1