Download (only when needed) the binary matching the version of Cypress installed
There are two issues with the current setup:
* we download the Cypress binary at each run even when no E2E tests are
  executed. We are downloading ~100MB on each CI run for most of the
  time nothing.
* the binary installed in the Docker image used to run the tests might
  be not match the installed version of Cypress. This might not end well
  in the future and it's not going to be funny to troubleshot.

This contribution disables the download of the Cypress binary when
building the Tuleap sources for packaging/automated test purposes
(basically what the CI pipelines do). Also, we make sure the correct
version of the Cypress binary is installed in the Docker image used to
run the tests. For developpers, the image will still be cached locally
avoiding an annoying download each time you want to run the E2E tests
but the cache will be busted when the version of Cypress changes in the
packege-lock.json (so no joke like "ooops I forgot the remove/rebuild
the image").

Part of request #14865: cypress typescript support

Change-Id: I7a1f19f76badbcaa30e12d35011ece18266baa86