Add error state illustration
If an error happen during the load of backlog items, then we should display an
illustration.

It also fix error illustration for campaigns when the theme is different
than blue (the crack should follow the theme color and not stay in
blue).

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: Ia54d2452af8dd1a09bc8dba754974d37e927ff70