Mark as expand/collapse the backlog item
Nothing really fancy, this only changes the caret to show the current
state for each backlog item.

This is a preliminary step to be able to see the tests of a backlog item
(in a next episode).

User can open the backlog item in a new tab.

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: I3ef2487b0bcf6c2e71583a49d39f1d0b9ccb9a1b