Display loading state when expanding backlog item
When a backlog item is expanded in the test plan view, we know display a
test definition skeleton.

This is another preliminary step to be able to see the tests of a
backlog item (in a next episode, stay tuned).

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: Ia56e9f730bd2a8ac6c2b78419cb45efbab260e41