Backend should enforce that the planning_tracker_id is linked to an existing and valid tracker
part of request #14986 report can throw a fatal error when db is inconsistent

Actually when user browse a tracker report, we will build a list of criterias.
Criterias might depends on external plugins (in our cas agiledashbaord).
When AD configuration is broken (planning_tracker_id = 0), then we have a fatal error on report.

This patch do NOT solve the fatal,
for now I only introduce a check to be sure that the value given to planning_tracker_id correspond to an existing tracker.

I will do a forgeupgrade to remove incorrect planning configuration in a dedicated changeset (and that will solves the fatal error)

Change-Id: I081be908ca6be14f9dfc830b6db90c27f4eb2e37