Better alignment of illustrations
If an illustration (error or empty) is displayed on the left (backlog
items) and an illustration (error or empty) is displayed on the right
(campaigns), then the ground of both illustration should be aligned.

For example if you have a new campaign, without any content nor
campaigns, then empty state should be on the same level.

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: Ie0a36d6a93cf933cf3300de5ecea6fe0b1a6bbe5