request #14991: Users might be able to push an LFS object to a repository migrated to Gerrit
Repositories migrated to Gerrit are supposed to be read-only for
everyone except the replication user.

Writing an LFS object to repository is supposed to be only possible for
users with a write access to it. This is not correctly verified for now,
users that are in groups that were allowed write access before the
migration are still able to push LFS object to this repository.
Repository that were using fine grained permissions before the migration
are not impacted.

Change-Id: Ic48c474bab32b21e6dd264d2ae076727c8fe8216