Autoscroll to [Add test] button
When you are redirected to the test plan after having created a test,
then the viewport autoscroll so that the [Add test] button is visible.

Let's be honest, this is not very terrific. And we are wrong since the
writing of the acceptance criteria. This solution (with switch to
tracker + redirection + autoexpand + autohighlight) is present only to
mitigate the cost of introducing a new artifact modal. It is introduced
as a temporary solution. As such, some compromises have been taken. The
autoscroll is clunky due to various reasons:
- We have to defer the scroll to let Vue finish creating the element
- Due to recursive load of tests, the ideal position may change during
time and thus may differ from our calculation
- Having the window to autoscroll while the user is propably interacting
with the page is not very UX friendly

Those are considered known issues and will be fixed when the modal to
create tests will be introduced.

Part of story #14888: Create a new test with artifact view

Change-Id: I8842fd50da008a4a2349f10e5324c8916bc4371a