Add the test status in the representation of the test definitions linked to a backlog item
The test statuses of the test definitions are retrieved in one SQL
query. This is more or less a mandatory requirement to not have horrific
performances with a real dataset. Indeed, the amount of data to retrieve
and filter is not small: all the test executions of the campaigns needs
to be intersected with the most recently updated test execution
associated with a test definition of a backlog item.

The SQL query has room for improvement but the subqueries are not
correlated which should be enough to keep the performance of the query
decent. Further improvements could be done once we have observed how the
query behaves with real production data. However, as a convervative
measure, the maximum number of test definitions retrieved at once via
the REST endpoint has been lowered.

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: I1099bdd123e29e5cfabe87b46d2bec979ee89786