Lower the max page size of the GET milestone/:id/testplan REST endpoint
With our current knowledge on how to retrieve the test status for each
test definitions we know that retrieving the test status of each backlog
items (story #14887) might be costful.
As lowering the max pagination size might be considered as a BC we
should do the change as soon as possible and relax the limit afterward.

Part of story #14884: Display list of backlog items with their tests definition

Change-Id: I7dd1e7576c1a36f363f0f788eba69a9184ac7813