request #15007: BC in the response of the REST endpoint GET /projects/:id/testmanagement_campaigns
The id field was indicated as int in the representation it was set as a
string. Changing the type of the property is a breaking change as it
breaks code relying on this endpoint.

Issue introduced by git #tuleap/stable/a9f9bb8f3fbcf18aad34720f545b95e539b3095d.

Change-Id: I5ddf7ebabba4179a542b73459e29cc54d6b78998