request #15013: Remove the ability to retrieve list of trackers via Ajax
Until recently, we were using an ajax route to retrieve the list of
trackers of a given project as an array of selectbox' options. This was
used in the legacy interface to create trackers (used to retrieve
templates in a project). Since this interface is no more available, we
can remove this dead code.

Change-Id: Idf45cc195192d70b18382c1897a81ba177c59857