request #15010: SSH keys must be sent to Gerrit without any kind of escaping
This seems to be the case for the whole 3.x branch, it is mentioned in
the documentation [0] and ultimately it makes sense since the content
type is set to text/plain.

While troubleshooting, we have seen that Gerrit was returning a 201
status code on success. This contribution also updates that but the
current behavior with a success on a 200 status code is also kept has it
the scenario described in the documentation.

[0] https://gerrit-review.googlesource.com/Documentation/rest-api-accounts.html#add-ssh-key

Change-Id: I7dc6bc14589c8d1364c28bcbdcedf881f5dc5ffc