Introduce Markdown in text fields
Note: This only introduces markdown rendering. Nothing is done on the
submit/edition of the artifact.

Go to an artifact.
Enter markdown content in a text field (in text format). Submit and
stay.
=> the markdown content is not interpreted.

Hack the database and for the corresponding entry in
tracker_changeset_value_text switch the format from text to markdown.
Refresh the artifact.
=> the markdown content is now interpreted.
=> edit the field, the field shows that it is in text mode. (and you
will loose markdown format on save).

Part of request #15026: Markdown support in trackers

Change-Id: I049a3e38688ad264a5f43342795158f6eeaff016