Introduce Markdown in follow-up comments
Note: This only introduces markdown rendering. Nothing is done on the
submit/edition of the artifact or its comment.

Go to an artifact.
Enter markdown content in a follow up comment (in text format). Submit
and stay.
=> the markdown cont is not interpreted

Hack the database and for the corresponding entry in
tracker_changeset_comment switch the format from text to markdown.
Refresh the artifact.
=> the markdown content is now interpreted.
=> If you try to edit the field, the formatting is lost. Since this is
not available to end users, the behavior is considered ok and will be
managed in a dedicated commit.

Part of request #15026: Markdown support in trackers

Change-Id: I36e9fcfd31cf023d0ab5b428005c1916b53dafec